// src/app/state/store/enum/enum.ts
export enum AdminActionTypes{
  LOAD_ORDERS = '[ORDERS] Load ORDERS',
  LOAD_ORDERS_SUCCESS = '[ORDERS] Load ORDERS Success',
  LOAD_ORDERS_FAILURE = '[ORDERS] Load ORDERS Failure',

  LOAD_PRODUCTS = '[PRODUCTS] Load PRODUCTS',
  LOAD_PRODUCTS_SUCCESS = '[PRODUCTS] Load PRODUCTS Success',
  LOAD_PRODUCTS_FAILURE = '[PRODUCTS] Load PRODUCTS Failure',

  LOAD_CUSTOMERS = '[CUSTOMERS] Load CUSTOMERS',
  LOAD_CUSTOMERS_SUCCESS = '[CUSTOMERS] Load CUSTOMERS Success',
  LOAD_CUSTOMERS_FAILURE = '[CUSTOMERS] Load CUSTOMERS Failure',

  LOAD_ORDERDETAILS = '[ORDERS] Load ORDERDETAILS',
  LOAD_ORDERDETAILS_SUCCESS = '[ORDERS] Load ORDERDETAILS Success',
  LOAD_ORDERDETAILS_FAILURE = '[ORDERS] Load ORDERDETAILS Failure',

  DELETE_ORDERS = '[ORDERS] Remove Orders',
  DELETE_ORDERS_SUCCESS = '[ORDERS] Remove Orders Success',
  DELETE_ORDERS_FAILURE = '[ORDERS] Remove Orders Failure',

  ADD_COMPLETED_ORDERS = '[COMPLETED ORDERS] Add Orders',
  ADD_COMPLETED_ORDERS_SUCCESS = '[COMPLETED ORDERS] Add Orders Success',
  ADD_COMPLETED_ORDERS_FAILURE = '[COMPLETED ORDERS] Add Orders Failure',

  LOAD_COMPLETED_ORDERS = '[COMPLETED ORDERS] Add Orders',
  LOAD_COMPLETED_ORDERS_SUCCESS = '[COMPLETED ORDERS] Add Orders Success',
  LOAD_COMPLETED_ORDERS_FAILURE = '[COMPLETED ORDERS] Add Orders Failure'
}